int/builder/markdown/hook: moves the link testing inside hook subpackage to test the function directly without file IO
